requirements.txt
allure-pytest>=2.13,<3
pytest>=8.3,<9
requests>=2.32,<3
selenium>=4.28,<5
